add agumentation metadata and some utils

+++ submodule_utils/tests/conftest.py
@@ -3,6 +3,8 @@ import pytest
 
 from submodule_utils.tests import OUTPUT_DIR
 
+CLEAN_AFTER_RUN=False
+
 @pytest.fixture
 def output_dir():
     """Get the directory to save test outputs. Cleans the output directory before and after each test.
@@ -10,5 +12,6 @@ def output_dir():
     for file in os.listdir(OUTPUT_DIR):
         os.unlink(os.path.join(OUTPUT_DIR, file))
     yield OUTPUT_DIR
-    for file in os.listdir(OUTPUT_DIR):
-        os.unlink(os.path.join(OUTPUT_DIR, file))
+    if CLEAN_AFTER_RUN:
+        for file in os.listdir(OUTPUT_DIR):
+            os.unlink(os.path.join(OUTPUT_DIR, file))
